refactor: extract test helpers into separate files
