+++ Vcr.HttpRecorder.AspNetCore.Tests/ConcurrentContextWebApplicationFactoryTests.cs
@@ -14,72 +14,6 @@ namespace Vcr.HttpRecorder.AspNetCore.Tests;
 public class ConcurrentContextWebApplicationFactoryTests
 {
     /// <summary>
-    /// In-memory interaction repository used to avoid file I/O in tests.
-    /// </summary>
-    private sealed class InMemoryInteractionRepository : IInteractionRepository
-    {
-        private Interaction? _stored;
-
-        public Task<bool> ExistsAsync(string interactionName, CancellationToken cancellationToken = default)
-            => Task.FromResult(_stored != null);
-
-        public Task<Interaction> LoadAsync(string interactionName, CancellationToken cancellationToken = default)
-            => Task.FromResult(_stored ?? throw new InvalidOperationException("No interaction stored"));
-
-        public Task<Interaction> StoreAsync(Interaction interaction, CancellationToken cancellationToken = default)
-        {
-            _stored = interaction;
-            return Task.FromResult(interaction);
-        }
-    }
-
-    /// <summary>
-    /// Helper class that manages an external API test server and provides its handler and base address.
-    /// </summary>
-    private sealed class ExternalApiFixture : IDisposable
-    {
-        public TestServer Server { get; }
-        public HttpMessageHandler Handler { get; }
-        public Uri BaseAddress { get; }
-
-        public ExternalApiFixture()
-        {
-            var host = new HostBuilder()
-                .ConfigureWebHost(webHostBuilder =>
-                {
-                    webHostBuilder
-                        .UseTestServer()
-                        .Configure(app =>
-                        {
-                            app.Run(async context =>
-                            {
-                                if (context.Request.Path == "/hello")
-                                {
-                                    await context.Response.WriteAsync("external");
-                                }
-                                else
-                                {
-                                    context.Response.StatusCode = 404;
-                                }
-                            });
-                        });
-                })
-                .Build();
-
-            host.Start();
-
-            Server = host.GetTestServer();
-            Handler = Server.CreateHandler();
-            BaseAddress = Server.BaseAddress;
-        }
-
-        public void Dispose()
-        {
-            Server?.Dispose();
-        }
-    }
-
-    /// <summary>
     /// Builds a WebApplicationFactory for the test's main server, adding VCR support
     /// and configuring an external API handler.
     /// </summary>
